Moved some stac function into utils module.
One of these functions is used in pxapi, and needs to be imported from a module not dependent on tensorflow.
