+++ pixels/stac.py
@@ -19,6 +19,7 @@ from pystac import STAC_IO
 from pixels import stac_training
 from pixels.exceptions import PixelsException, TrainingDataParseError
 from pixels.mosaic import pixel_stack
+from pixels.stac_utils import stac_s3_read_method, stac_s3_write_method
 from pixels.utils import write_raster
 
 # Get logger
@@ -185,18 +186,6 @@ def upload_files_s3(path, file_type=".json", delete_folder=True):
         shutil.rmtree(sta)
 
 
-def stac_s3_read_method(uri):
-    parsed = urlparse(uri)
-    if parsed.scheme == "s3":
-        bucket = parsed.netloc
-        key = parsed.path[1:]
-        s3 = boto3.resource("s3")
-        obj = s3.Object(bucket, key)
-        return obj.get()["Body"].read().decode("utf-8")
-    else:
-        return STAC_IO.default_read_text_method(uri)
-
-
 def list_files_in_s3(uri, filetype="tif"):
     parsed = urlparse(uri)
     if parsed.scheme == "s3":
@@ -223,31 +212,6 @@ def list_files_in_folder(uri, filetype="tif"):
         return list_files_in_s3(uri, filetype=filetype)
     else:
         return glob.glob(f"{uri}/**{filetype}", recursive=True)
-
-
-def stac_s3_write_method(uri, txt):
-    parsed = urlparse(uri)
-    if parsed.scheme == "s3":
-        bucket = parsed.netloc
-        key = parsed.path[1:]
-        s3 = boto3.resource("s3")
-        s3.Object(bucket, key).put(Body=txt)
-    else:
-        STAC_IO.default_write_text_method(uri, txt)
-
-
-def get_catalog_length(catalog_path):
-    if catalog_path.startswith("s3"):
-        STAC_IO.read_text_method = stac_s3_read_method
-        STAC_IO.write_text_method = stac_s3_write_method
-    try:
-        collection = pystac.Collection.from_file(catalog_path)
-        size = len(collection.get_child_links())
-    except Exception as e:
-        sentry_sdk.capture_exception(e)
-        catalog = pystac.Catalog.from_file(catalog_path)
-        size = len(catalog.get_item_links())
-    return size
 
 
 def parse_prediction_area(
